see branch's new name
switched variable sfor spells to hashtables though I may revert this as I wanted more  a database for the spells and this accomplishes nothing regular array lists didn't.

Camera tracking to four different places, including the PC's seaside home for the tutorial and later in the game, as well as having designed that home.