bug fixes, sorting layer swapping fixes/changes, and spell overhauling continue
Fixed some portal points, fixed sorting layer swapping wrt the proper scripts but it needs way more testing, continue to overhaul the spell system including reverting some changes because hashtables are pointless for this. I hate bugfixing, so I have been dodging it and now there is sooo much of it to be done. More later.