tons of changes since I last visited this branch.
Varies from the AnxietyGame branch now because I've done some bug fixes, changed how one gets to the ruins based on story we've added on to while also requiring the player to interact with at least a few NPCs.

Camera dimensions have changed. ObjectInteraction script is nearly empty but has child scripts that do different things while allowing player interactions to affect them through the input system with the same command.

I am in the process of reorganizing the files.

Dealing with an occasional memory leak that I have documented on the Discord server.

Will make the readme file more informative in the future, as well.